The login now uses basic auth to see if the username and password are correct
Also added a Api file that the app can use to make calls to the site's rest api.
For it to work on the csci server you need to alter the android emulator's host file.

Need to add a few more touches to the login page to catch any exceptions there may be.